Correciones

+++ src/main/java/com/lenguageconquers/service/serviceImplement/AvatarServiceImpl.java
@@ -1,6 +1,7 @@
 package com.lenguageconquers.service.serviceImplement;
 
 import com.lenguageconquers.dao.AvatarDAO;
+import com.lenguageconquers.dao.EstudianteDAO;
 import com.lenguageconquers.model.Avatar;
 import com.lenguageconquers.model.dto.AvatarDTO;
 import com.lenguageconquers.service.AvatarService;
@@ -19,6 +20,9 @@ public class AvatarServiceImpl implements AvatarService {
 
     @Autowired
     private AvatarDAO avatarDAO;
+
+    @Autowired
+    private EstudianteDAO estudianteDAO;
 
     @Override
     public String registrar(AvatarDTO avatarDTO) throws Exception {
@@ -39,25 +43,30 @@ public class AvatarServiceImpl implements AvatarService {
         return "Se creo el avatar exitosamente";
     }
 
-    //TODO:NO ESTA COMPLETO
     @Override
-    public Avatar actualizar(Avatar avatar) {
-            return avatarDAO.save(avatar);
+    public Avatar actualizar(Avatar avatar) throws Exception {
+        if(!Validaciones.validExtensionImg(avatar.getImgAvatar())){
+            throw new Exception("Debe ingresar una extension valida para la imagen");
+        }
+        return avatarDAO.save(avatar);
     }
-    //TODO:NO ESTA COMPLETO
+
     @Override
-    public void eliminar(Long idAvatar){
+    public void eliminar(Long idAvatar) throws Exception {
+        if(idAvatar == null){
+            throw new Exception("El id del avatar es obligatorio");
+        }
+        if(avatarDAO.existsById(idAvatar) == false){
+            throw new Exception("No se encontro el avatar con ese id");
+        }
+        if(!estudianteDAO.findByIdAvatar(idAvatar).isEmpty()){
+            throw new Exception("No se puede eliminar el avatar porque esta siendo uilizado por un estudiante");
+        }
+
         avatarDAO.deleteById(idAvatar);
     }
 
-    /**
-    //TODO: VERIFICAR METODO
-    @Override
-    public Avatar listarId(int id) {
-        return null;
-        //return avatarDAO.findOne(id);
-    }
-*/
+
     @Override
     public List<Avatar> listar() {
         List<Avatar> avatares = avatarDAO.findAll();
